LTX: Added longturn more experimental ruleset based on LTT
LTX is not an experimental ruleset but something with more experimental features. The goal with LTX is to have a game with features familiar to veteran players while adding new ways to do things. The original LTT ruleset will get bug fixes and features decreasing the importance of tc actions while more liberal stuff is added to LTX. This will allow LTT playes to play with more stable ruleset while LTX is for those who want to try... more experimental stuff.