LTX: Port of Guangzhou cost 25. no upkeep (was 60 and upkeep 2)
Having this to add to the bonus with the lighthouse might be too much. Instead it's now an alternative to the lighthouse. You can build the ligthouse and have the move bonus and veteran units or you can have the cheap port wonder making the ships veteran but without the move bonus.