LTX: Terracotta Army GreatWonder +1 v level
Terracotta Army is aGreat Wonder becoming available with construction. It gives the units built in the city one extra veteran level. The players can now choose if they should race for construction to get one city to build v, vv or even vvv units or if it makes more sense to start more early with Sun Tzu's. Being a great wonder it may also make sense to get it for defensive purposes. The enemy can't build it if it's already built.