LTX: Creating new land on ocean no longer needs adjacent tiles
This change needs to be evaluated at a later time. Needing no adjacent tiles for transforming ocean into swamp allows the players to create artificial islands on the sea if shallow ocean is available. Creating bases on strategic locations makes it much more important to gain naval supremacy and adds another reason for monitoring the activities on the oceans.