LTX: cities can be poisoned if they are bigger than size 8
Poisoning cities to size 1 is too powerful and most of the time effectively makes the city irrelevant for the rest of the game. Poisoning to size 8 removes the defensive bonus the cities get at size 9. This way it's still possible to recover and keep the border area cities reasonably relevant. This also allows the smaller nations to have more than a few cities that remain productive.