LTX: Manhattan project cost 800 (was 2500)
The big nations no longer want or need nukes if it's already possible to win with the production capacity. The smaller nations can't afford spending 2500 to the nukes but they might be able to afford 800. The big nations can't afford making all the cities nuke proof and this may make it a big issue to those who keep big stacks of units in the cities when they attack the smaller nations.