LTX: Siege weapons get 60% healing from barracks
These units are not made veteran if built in a city with barracks but can take some advantage from them. The players are able to heal faster but not as fast as with other units. This slows down attacks with siege weapons but not as much as before. This way the enemy may have a better chance to retaliate one turn before the attack with the siege weapons continues.