LTX: Warriors belong to ancient land unit class
Allows the players to send some kind of military presence to far away lands. This is not supposed to allow big military operations but make it less safe to keep the coastal cities empty or virtually undefended. This also allows placing low end military units to far away locations before better ships are invented. It's also intentional that the players may lose the ability to send military units overseas if the player doesn't plan the new techs properly.