LTX: Ships no longer cause unhappiness when out
The purpose of the ships is to be out on the sea. This is why it doesn't make sense to make the citizens back home unhappy. From the game point of view this makes it more easy to have units on the oceans and the game needs more action for the sea units.