LTX: adjusted unhappiness for conquered cities
No unhappiness for anarchy since there is no system anyway. Communism gets less unhappiness so that it's god for conquering while not that great for research or economy. Federation is supposed to be equally good to democracy and that's why it gets the same penalty for conquering cities. City states has a minor penalty.