LTX: Fanatics A4 (3) D5 (4) cost 30 (35)
Also making fundamentalism more powerful. There is a gap for funda in the mid game but this is intentional. It can't be used at all times but the intention is to make it useful in the late early game and early late game. id game and late game are maybe not that good for fundamentalism.