LTX: terrain changes fixed
Making forest into plains is not that hard but making it into grass is much harder while possible. Swamp into grass is harder but not really hard- Plains is hard to turn into forest. This makes the grass strategy less powerful but allows using it if the player is focusing on worker production.