Royale: Base ruleset LTT
Royale will be the updated version of LTT. It's not extending the game as much as LTX does. The design goals are currently not fully defined and there may be need to iterate while the ruleset is improved. There should be less iteration compared to LTX. LTT will remain stable as it is. The goal is to update Freeciv21 Royale with this version once it's good enough.