Royale: Upgrading 2 unist for all players with Amph War (radio)
Getting marines is a winning tech and making it to upgrade 2 units / turn for everyone makes it slightly more fair to those player  who doesn't have the tech. The upgrade feature was moved from radio since the games are too often decided before anyone knows radio tech. Too often from the point of view of upgrading the units for balancing the game.