Royale: No city poisoning if the size is below 17 (was 2)
The smaller nations are virtually wiped out in only few turns if city poisoning is allowed with no limits. Preventing to poison cities with size 16 or below keep the front line cities viable. However if the cities are much bigger its still possible to poison and this way hurt players with exposed cities. This also makes the game more playable for those who are lagging with techs. The first player with spies can no longer dominate the front line with spies only. This also makes it more reasonable to build colonies or trying to improve the front line areas. Those spots are no longer considered a lost cause once the spy wars begin.