Royale: Unhappiness because of the foreign citizens
Makes conquering the enemies less easy by making the foreign citizens to result unhappiness. Once the citizens convert to the same as the owner of the city, the effect is reduced and eventually removed. It's unrealistic to have all the resources of a conquered city with no unhappiness. From a gameplay point of view no unhappiness also gives the player too big of a bonus while also allowing defeated players to fully revenge to their enemies by allowing the enemy of the enemy to conquer all of the cities. This also allows strategies where the conquering player builds migrants for workers and then maybe grows the city back. 30% of the foreign citizens will be unhappy. Except for communism where this number of 40%. With fundamentalism only 10% of the people are unhappy (unhappy because they are of foreign citizenship).