Royale: The explorer no longer ignores Zones of Control
The explorer is traditionally a very powerful unit able to break Zones of Control in the early game. It's also able to investigate the enemy lands very effectively since the other units can't block it with the ZoC. Removing the ZoC makes it more like an ordinary unit and makes the game more easy for those who use military units for blocking the enemy actions. Breaking the ZoC is still possible with diplomats, airplanes and similar units but this makes the ZoC breaking considerably more difficult.