Royale: Great Library gives 50% bonus without libraries
Making the Great Library to help everyone makes it a tough choice for the player building it. The bonus for the player building it takes a long time to match the cost of the wonder. This will also help the less developed nations to keep up with the more advanced ones.