Royale: Internet adds 50% bonus to the Universities
Adding the bonus to Universities instead of Research Labs allows the less developed nations to have a chance with catching up. If the bonus would only apply to the research labs only the most advanced players would benefit from it. This is another step for boosting the existing buildings most of the players have had to time to build.