Royale: Empire sizes are halved
Smaller empire sizes result with less stuff to manage and this is especially important for the late game. Previously late game has been consuming great amounts of time to manage and this is an attempt to fix that. This also makes it more easy to set up maps where competing about space for the cities is not the main objective. Instead it's more easy to make setups where the players are able to build empires gradually expanding towards each other. This kind of setup would be the opposite to grabbing all the possible land as soon as possible while making the game less about the skill to expand really fast in the early game.