Royale: Howitzers have 7 moves (was 9)
This change allows the players to build defensive terrains protecting the cities from being easily overrun with howitzer units. It's still possible to take cities with these units but together with the reduced empire sizes the battles need to become more strategic. There may be more need to build fortifications close to enemy cities. This change also attempts to bring out the battles from the cities. This would be the opposite to the traditional city to city hopping where the units can spend almost all the time inside the cities.