Royale: Nuclear units available with Steel and Radio
The nuclear units have 9 moves on default and will get 9 more from advanced flight. This change makes building the manhattan project a huge gamble since most nations may be able to build nuclear weapons. The first unique nuclear weapon will create a huge risk to single cities close to the enemy. Moving the nuclear units with ships requires later techs and this greatly limits the usability of the weapon in the hands of less developed nations. However a single nuke may allow a less developed nation to create a chaos on the front lines.