Royale: Hanging Gardens affects one or two cities
The old main effect of the wonder will be moved to the empire base size. Now this wonder makes 3 people content in the city it's located. On top of that it will make two people happy in the capital city but only if it's not located there. This effect allows the players to decide how they want to divide the effect between two cities. This also creates a situation where the players may need to have better defense for two cities instead of one with the wonders.