Royale: Doubled the base empire sizes
The main effect of the hanging gardens was moved to base empire sizes. This allows the players to grab a reasonable amount of land in the early game without building a wonder that would be a must to build one. Not building hanging gardens is also something that really hurts the less experienced. players. While the game is about skill this ruleset makes some things more easy for the less experienced ones while replacing the effects with new but slightly less important features.