Royale: Marco Polo smalle bonus and specials give bonuses at different times
It's less about getting the instant bonus from Marco Polo and more about it unlocking the ability to boost the additional bonus from some specials. The unlocking for all the bonuses from the specials happens when any of the players discovers the required tech or builds the building. This may also reveal if someone has found some particular tech.