Royale: Ivention upgrades 2 units each turn and electricity 1
Progress is inevitable and keeping the unit more up to date makes the game more easy for the less experienced players. It's also boring to fight against those who have no resources to upgrade the obsoleted units. This also makes the players better in defense since the obsoleted units may be auto upgraded. It's more hard to conquer those who struggle with defenses. This should also make it less likely for the expansionist players to grab lots of land in the mid game. Doing that is supposed to be more hard on Royale but maybe not on other rulesets.