Royale: Unescorted howitzers can be captured and no citybuster
This should make it more easy to support other nations by allowing then to capture some units. It's will also add more need to actually defend this offensive unit. The ciede units no longer are citybuster units and this will make it less easy to take cities with this unit class. Building veteran units compensates some of that but not all. The goal is to make it less easy to conquer entire nations in only a few turns.