Royale: Volunteer Coprs and Infantry Legion units
Mercenary versions of the riflemen and infantry units. Very high upkeep but no happiness penalty if away from home. They can however be captured if not guarded properly. Mercenary units are not the most loyal ones. This unit allows the players to give defensive units as support to other players since capturing mercenaries makes this possible.