Royale: Literacy and University both upgrade one unit every turn
Players lagging on techs often need to rush for upgrades. This results in more trouble in keeping up with the latest techs. It also makes no sense for a more modern world to keep supporting obsolete units if they are eventually not upgraded to modern ones. More upgrades also make it less devastating if the player is forced to focus on building defensive units. This also benefits attackers, since the attacker will get modern units much sooner. This should also add a new strategy to the game, since players need to figure out whether building a new unit or a building is more profitable just before the units start to be upgraded.