Royale: Forts and pollution result with penalty to a city
If a city has pollution in the working area it will get -20% penalty to the shield production. This results with immediate need to clean up the pollution unless the player is fine with less production. If the city has a fort in the working area it will get -20% penalty to trade. This can be used as a tool for making your enemy less powerful. Build a fort next to production cities and the enemy suffers without any killed units. This should create new options for nations bordering each other to annoy the other one since both effects work even if the special is on foreign land.