Royale: Added Silk Road (Trade) Gold road moved to Banking
Silk Roads fits better to Trade and since it's obsoleted with navigation there is a reason to delay getting navigation if the player has silk and it's not that important to get new ships. If the new ships are important en obsoleting this is a trade off for advancing the nation.